requirements.txt
numpy
torch
torchvision
torchaudio
pyyaml
pandas
matplotlib
